feat/1193 - bunch of small refactors and cleanups
