+++ src/components/CohortBuilder/Table/index.js
@@ -3,60 +3,28 @@ import { compose } from 'recompose';
 import { withTheme } from 'emotion-theming';
 
 import { withApi } from 'services/api';
-import { demographicQuery } from './DemographicTable';
+import { participantsQuery } from './queries';
 
-import BaseDataTable from 'uikit/DataTable';
 import QueriesResolver from '../QueriesResolver';
-
-const participantsTableViewColumns = [
-  { Header: 'Participant ID', accessor: 'participantId' },
-  { Header: 'Study Name', accessor: 'studyName' },
-  { Header: 'Proband', accessor: 'isProband' },
-  { Header: 'Vital Status', accessor: 'vitalStatus' },
-  { Header: 'Diagnosis Category', accessor: 'diagnosisCategories' },
-  { Header: 'Diagnosis', accessor: 'diagnosis' },
-  { Header: 'Age at Diagnosis', accessor: 'ageAtDiagnosis' },
-  { Header: 'Gender', accessor: 'gender' },
-  { Header: 'Family ID', accessor: 'familyId' },
-  { Header: 'Family Composition', accessor: 'familyCompositions' },
-  { Header: 'Files', accessor: 'filesCount' },
-];
-
-const stringListView = data => (
-  <div>
-    {data.map((datum, index) => (
-      <div key={index}>{datum}</div>
-    ))}
-  </div>
-);
+import ParticipantsTable from './ParticipantsTable';
 
 const enhance = compose(
   withApi,
   withTheme,
 );
 
-const Table = ({ theme, sqon, api }) => {
+const ParticipantsTableView = ({ sqon, api }) => {
   return (
-    <QueriesResolver api={api} sqon={sqon} queries={[demographicQuery({ ...{ sqon } })]}>
-      {({ isLoading, data, error }) => {
-        const formatters = {
-          diagnosisCategories: stringListView,
-          diagnosis: stringListView,
-          ageAtDiagnosis: stringListView,
-          familyCompositions: stringListView,
-        };
-
-          return (
-          <BaseDataTable
-            columns={participantsTableViewColumns}
-            data={data}
-            transforms={formatters}
-            loading={isLoading}
-          />
-          );
-      }}
+    <QueriesResolver api={api} sqon={sqon} queries={[participantsQuery({ ...{ sqon } })]}>
+      {({ isLoading, data, error }) =>
+        error ? (
+          <div>{`${error.name} - ${error.message}`}</div>
+        ) : (
+          <ParticipantsTable data={data} isLoading={isLoading} />
+        )
+      }
     </QueriesResolver>
   );
 };
 
-export default enhance(Table);
+export default enhance(ParticipantsTableView);
